Move validate function to the abstract validator.
